requirements.txt
clip @ git+https://github.com/openai/CLIP.git
dask==2023.5.0
diffusers==0.26.2
evaluate==0.4.1
fairscale==0.4.13
huggingface_hub==0.20.3
img2dataset==1.43.0
nltk==3.8.1
numpy==1.24.4
open_clip_torch==2.23.0
openai==1.13.3
opencv_python_headless==4.9.0.80
pandas==1.5.3
Pillow==10.1.0
pycocoevalcap==1.2
pycocotools==2.0.7
Requests==2.31.0
retriv==0.2.3
ruamel.base==1.0.0
SceneGraphParser==0.1.0
scikit_learn==1.3.2
scipy==1.12.0
setuptools==68.0.0
spacy==3.7.2
stanza==1.6.1
timm==0.9.16
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tqdm==4.66.1
transformers==4.37.2
imagededup==0.3.2
